Allow opening of SAM/BAM files by fd
Opening by fd will allow us to use pipes so that more generic
communication can be performed. It also allows us to disentangle
ourselves from the use of magic filenames to get stdout and stdin
streams (this is retained for compatibilty within samtools, with the
hope that at some stage it might be moved to the application layer).

Documentation for samopen also updated to include all options and
correct an incorrect statement in the docs. It is not clear why the
'u' mode is provided since the default compression mode is 0 anyway.