Include htslib rules and add header dependencies
Use the rules and variables from htslib.mk to write out exact
dependencies on htslib headers and (for executables) library.
With this, samtools builds with "make" from here are correct
even when htslib source code is being modified.

(A little more infrastructure is needed to select between static
and shared libhts.)