Warn and use correct bin when bin field is wrong
The abort introduced in pull request #43 (8e3b3f2 "Verify old BIN value
of reads during 'samtools index'") is too Draconian, as this field is
in practice used only during indexing -- so the BAM file is mostly fine.

If we're recalculating it here anyway, we might as well use the corrected
value to generate the index and just warn (in a rate-limited way) about
the anomalous records.