Fix bin calculation for unmapped reads
Unmapped reads should have a CIGAR of "*", but sometimes do not e.g. as
occasionally emitted by bwa.  Our bin calculation has been changed to
base the bin on a 1-base region for reads that are either unmapped or
have no cigar, and now agrees with Picard's for these reads.