Reenable tview and port to htslib
tview: reenable the tview command by reverting removal of components essential to
tview.  Also port tview over to using the htslib header files.  Several new files
are created to house existing bits of code which were part of one huge lump but
needed to be split out when that lump was deleted.