Fix old API samopen() wrapper fn_aux bugs
For reading, the old samopen()'s third parameter is a FAI filename for
use if sequence headers are missing from the input file; for writing,
it is the headers that are to be written out.

It's not too clear what hts_open()'s third argument is for, but it's
certainly not for output headers.  So pass it as NULL when writing.

Also call sam_hdr_write() for the is_cram case too.