Remove incorrect parameter check
The q parameter is an m*m matrix, and is allocated by the caller and
m & q are passed in accordingly.  It appears that this check may have been
intended to be "if (m > qmax)" but we have no independent idea of the size
of q, so there's no way to validate m against it.

Values in q are computed from aux.bsum[k]/etc where 0 <= k < m, so m is
also limited by the size of bsum[], namely 16.  However meaningful indices
for bsum/etc mostly are 0...4 (i.e., A,C,G,T,N) and in practice (currently)
m is either 4 or 5 (with q sized accordingly).  So checking "if (m > 16)"
would be somewhat pointless.