Adds tests for samtools view.
Adds .sam, .fa, .fa.fai and .bed files as test data for the view command.
Adds tests in the test.pl script.  Exercises most options, apart from -B,
-@, -L, -t, -T and -s.  Most tests work by taking the test/dat/view.001.sam
and modifying it in the same way as samtools is expected to, and then
comparing the result.  BAM and CRAM conversions are tested by converting
back to SAM and checking that the output is equivalent to the original copy.