Adds tests for the new -x tag stripping option.
Includes tests for single and multiple -x arguments, and mixing -x with
-r (samtools issue #153: -r disables -x).  Also tries to filter by read
group and strip RG tags.  This currently doesn't work due to issue #153,
but should do once that has been fixed as long as tag stripping is still
done after the filtering stages.

Includes a few bugfixes to the tag stripping code in filter_sam.