Adds tests for the -s (subsampling) and -@ (threads) options
Generate a big sam file (1Mbp consensus).  This is compressed with bgzf
so it doesn't use up too much space.  It is then used to test the -s and -@
options, which require a lot of input data to do anything interesting.

-@ is tested by making bam and cram files with different threads settings
and checking to see if they are the same.  -@ doesn't actually do anything
at the moment, so this is in place waiting for when it does.

For -s, the file is filtered, and the results are checked to ensure that
all reads came back paired and that approximately the right number were
returned.  This has to be an approximate check due to the random nature
or the sampling operation.

The sam_compare function is adjusted slightly to allow for CRAM adding
M5 tags to @SQ lines without one.  The content of the M5 tag is not
currently checked.

The compare_sam option of run_view_test is updated so that it can read from
a samtools view pipe instead of writing a file and then comparing it.  This
is done to avoid writing lots of uncompressed copies of the big SAM file
when running the threads tests.