Makes filter_sam filter @RG lines when filtering by read group.
Adds code to strip unwanted @RG lines from the SAM header when read group
filtering is enabled.  Also modifies the library filter so that it does not
try to strip @RG lines.  It is also made to work in the same way as samtools
view when combined with read group filtering - i.e. reads that do
not pass both filters are removed.