Adds tests for 'samtools cat'
Function cat_sams simulates the action of 'samtools cat', and test_cat tests
it by comparing the output to that of cat_sams.  bgzip is used to decompress
and recompress the bam files to break the relationship between read and bgzf
block boundaries.  This is used to check that samtools cat can properly
handle the case where the header and some alignments share the same bgzf
block.

run_view_test gets a 'cmd' argument that allows it to run subcommands other
than 'view'.  This unfortunately means it is now not very well named.