Adds tests for samtools depad.
Depads a test SAM file, and compares the output with a version that has
been depadded by hand.  The files include comments describing the
padded and depadded CIGAR strings for each test read.

The test is currently a bit restricted.  It can only test the case where
the -T option is used due to some fundamental problems with updating header
lines when embedded references are used.  CRAM is not tested either as
the way embedded references are stored in SAM/BAM does not seem to fit well
with the CRAM specification.  Currently trying to convert to CRAM strips out
the CIGAR string of the reference, which removes the information about where
the pads are positioned in the reference.