Adds expected fail and unexpected pass.
Add code for handling expected failure and unexpected passes.  Change the
exit code to be non-zero if there are any unexpected failures or
unexpected passes.  If there are only expected passes and expected failures
then the exit code will be zero.  The rationale for failing on an
unexpected pass is that the test script needs to be updated to expect a
pass in that case.

Updates the samtools view -l and samtools depad tests to be expected failures.