Fixes an issue with -i0 involving truncation of insert sizes.
Setting '-i 0' previously caused any inserts of the maximum size to
be omitted. This is due to a bug at line 111 of stats_isize.c where
the maximum insert size is returned as the 'number of items', whereas
it should be the maximum plus one.

This should then cause test 6.stats to succeed. However, it still fails
due to a $/# typo in the expected results. This commit also therefore
fixes the expected results. The same issue caused 5.stats to fail, so
we fix that and remove the expect_failure there for good measure.