Silence -Wstrict-prototypes warnings
In C (but not C++), "int foo();" is a non-prototype that says nothing
about what parameters foo might or might not have.  To say that it has
no parameters and avoid warnings with the GCC option above, you need
to write "int foo(void);".  (In C++, both imply foo has no parameters.)

Fixes #45.