Fix getline memory leak in test_bam_translate
getline resizes *res with realloc if *len is smaller than
the number of characters read from the stream on that call.
Setting *len to 0 before each test thus causes getline to
use realloc on *res to increase its size by the number of
characters read for each test, despite the fact it is already
of the correct size.