Remove our bgzip.c in favour of htslib/bgzip
Now that bgzip is part of htslib, don't build a version of it here
and remove the now-outdated source code.

Parts of the test suite use bgzip, so add infrastructure to supply
an explicit bgzip path as required.  Eventually, there will be a
Makefile mode to use a system bgzip along with a system htslib; such
a mode will provide an empty $(BGZIP) and so "test.pl --exec bgzip="
invokes an executable named "bgzip" via $PATH.