Removes all randomness from `samtools merge`
The random number generation in `samtools merge` is unnecessary.
Random numbers in the IDs have been replaced by the index of the
input file from which each record comes.

This is (a) very very slightly faster and (b) does not require
one to supply a random seed just to get identical merge output
from two otherwise identical runs.

The drawback is when merging files which have themselves already
been merged, one can expect more collisions, and IDs that will
take the form of original-1-2-3-1-2, which may be more verbose
than desired. On the plus side, these are potentially usefully
informative as to where the record came from.