Improves performance of `samtools merge` (slightly)
Moves call to pretty_header so that it is done once at the end rather than
once for each input file. As a result, 'trans_tbl_init' no longer returns
prettified headers. This results in a performance improvement when merging
many files, although performance still becomes very poor when merging on
the order of hundreds of files.