Add seq_cache_populate.pl
seq_cache_populate.pl builds a cache of reference files for the CRAM
codec to use.  It can be given a list of fasta files to read on the
command line, or alternatively it can be made to search a directory
tree for anything that looks like a fasta file.

The sequences in the resulting cache are stored in files named after
the MD5 of the sequence itself.  Samtools/htslib can be made to use the
cache by appropriate setting of the REF_PATH environment varaiable.