Overhaul samtools sort options and invocation
New set of samtools sort options, with a more typical -o FILE option,
separate options for final output and temporary files, and the ability
to write SAM, BAM, or CRAM.  Fixes #171.

The old usage is deprecated but still present, and is assumed when
neither -O nor -T options are present.

Updated documentation to list all options, and mention that the old
usage is now old.