Replace bcftools+bcf_filter.sh with new vcf-miniview
Add new vcf-miniview test utility, which reads BCF or VCF and outputs
VCF optionally filtered in the same way that the old bcf_filter.sh did.
The sed -r option used by bcf_filter.sh was not available on OS X, and
an installation of bcftools will not necessarily be available on test
machines, notably when testing via travis.