samtools bam2qf option to force FASTA or FASTQ output
This addresses issue #313 - the current behaviour (left as the
default) produces a mixture of FASTQ and FASTA depending on if
each individual read has quality scores or not.

While some tools/parsers will handle mixed files like this, others
will have trouble. This is a problem with systems like Galaxy with
a strict datatype system.

I wanted to use -O for the new output format argument (to follow
the pattern set in samtools sort/rmdup/fixmate) but that is already
taken for the original quality option.

Note that if samtools bam2fq is called forcing FASTQ output, this
currently will use a default PHRED quality score one 1. This value
might make sense to add as a further command line option?