[Makefile] Use user's CPPFLAGS/LIBS variables
Remove INCLUDES and add EXTRA_CPPFLAGS for our own various options that
need to stay set even if the user overrides CPPFLAGS.  If we end up with
necessary compiler or linker options or libraries, we'll add EXTRA_CFLAGS,
EXTRA_LDFLAGS, or EXTRA_LIBS accordingly.

Rename LDLIBS to LIBS; it seems that LDLIBS doesn't particularly
exist outside GNU Make's built-in rules.  Instead LIBS is configure's
conventional variable for the purpose, so is the right thing to use
for a user-visible variable.

[NEWS]
* Samtools's Makefile now fully supports the standard convention of
  allowing CC/CPPFLAGS/CFLAGS/LDFLAGS/LIBS to be overridden as needed.
  Previously it listened to $(LDLIBS) instead; if you were overriding
  that, you should now override LIBS rather than LDLIBS.