Rationalised the long option handling for "global" options.
There is a new hts_open_opts too that takes an option structure.  This
is a change in behaviour as the former code used hts_open followed by
an hts_opt_apply call to apply the options to the file handle.

Many more sub-commands to-do still.  This is largely a patch for the
option infrastructure and a few subcommands to test it with.