Allow "cmd | samtools view" and write usage to stdout
Given just "samtools view", display the usage or view stdin (as per
"samtools view -") depending on whether the input is a terminal
similarly to other subcommands.  This makes samtools view act more
like a typical Unix filter in pipelines.  Hat tip @jwhsanger.

When it's not an error message (as in "samtools --help view" or just
"samtools view"), display the usage on stdout and exit successfully.