Switched to using htslib instead of bam.h.  Plus CRAM support.
Also had to fix the test data to not have unmapped reads with a CIGAR
string as this isn't supported in CRAM.  The change to stop this
blowing up in CRAM now also means that these strings do not get edited
in SAM or BAM either.  I think that is valid though as they are
currently nonsensical.