Adds `samtools quickcheck` subcommand
`samtools quickcheck` opens a set of SAM/BAM/CRAM files and
checks that the header of each contains at least one target
sequence, then (currently for BAM files only due to htslib
API limitations) seeks to the end of the file to check
whether the EOF block is intact.

Exits with a zero status if all input files checked out ok,
or a non-zero status otherwise. Can be made to give a list
of failing files (on stdout) using the `-v` option or to be
more verbose (on stderr) by giving additional `v` options.