Made samtools cat work, when possible, with CRAM files.
The cram variant is subtly different in that where possible it
reconstructs the joint header by stitching the @RG lines from the
concatenated files, but the -h option is still available.

Changed it so it works with 1 file.  It's a bit esoteric, but
technically there is no reason why it requires 2.  (Unix cat copes
with 1 just fine.)

Also fixed a small memory leak in reheader.