Handle malloc failures in bam_sort and cleanup exit calls
If a malloc fails in pretty_header, print an error and return true.
If a malloc fails in trans_tbl_init, print and error and die.
Replace abort() calls with exit(EXIT_FAILURE) and informative error message.
Replace exit(1) and exit(-1) with exit(EXIT_FAILURE).
Improve quality of some of the error messages when we fail.