Replace xz-compressed test data file with gzipped equivalent
Avoid xz, which may not be installed by default especially on OS X
or Windows.  To improve gzip compression of this data, we only gzip the
interesting fields and reconstruct chromosome/position in fields 1 and 2,
reducing the gzipped data from 3M to 582K -- hat tip @jkbonfield.

(The doubled \\ escapes the backslash from sh(1)'s read builtin.)