Improve header merging.
Add new merged_header_t struct to store state for the combined output
header.  This is passed in to trans_tbl_init instead of a bam_hdr_t.  It
includes hash tables which allow the presence of a @SQ SN, or @RG or @PG
ID in the output to be rapidly tested.  It also includes separate kstrings
for the various record types, which makes it easier to get the desired
ordering of records in the combined header.

The monolithic trans_tbl_init function is broken up into smaller parts that
handle the different record types, and rewritten to use the new
merged_header_t struct.  A couple of combined functions deal with the RG
and PG records, eliminating some duplicated code.

pretty_header, regex_escape and copy_headers are removed as they are no
longer needed.

New functions init_merged_header and free_merged_header are added to create
and tear doen the merged_header_t struct, along with finish_merged_header
which creates a bam_hdr_t from the data in the merged_header_t.  This
effectively replaces pretty_header.

bam_translate prints warnings to stderr about tags that do not appear in
the header.  It now tracks which ones it has warned about so it only
prints one message for each combination of ID and input file.

test_trans_tbl_init.c is modified to work with the new merged_header_t
struct.  A lot of duplicated code is also eliminated, especially from
the setup_test_ functions.

As pretty_header has gone, test_pretty_header is no longer needed and also
removed.

test/merge/3.merge.expected.err is changed to reflect the changed
warning message from bam_translate.