Added -o/-U filename.<suffix> detection to samtools view.
Also fixed a bug where using -C and -U together failed due to not
specifying the reference to the -U cram file.

The suffix used can differ between -o and -U.  The format is
overridden by -b / -c or --output-fmt.  It is not possible to specify
a different format for -o and -U except via suffix guessing (this was
not possible previously either as there is only a single -b/u/C
option).