Make merge allow input files with no @HD line
Remove check added in 55c852f that required the first SAM header to have an
@HD line when merging.  Previous versions worked if no @HD lines were
present; this allows the new header merge code to do so as well.

The new behaviour is that the first @HD line found on either the file
given in the '-h' option, or any of the input files, will be used.  If no
@HD line is found the program will complain, but it will continue and
exit successfully (assuming it otherwise worked).

Some code is included to add an @HD if none was present on the inputs, but
it's currently commented out pending discussion.

Fixes #475