Added bz2 and lzma to Makefile (mandatory).
If people want to avoid requiring these, eg building samtools without
CRAM support, they'll need to use the configure script.

TODO: Get configure script working via some form of pkg-config
approach so it knows which libraries to link against.  This is a
larger issue (it doesn't work for curl or irods either) and so
not addressed in this PR.