Tidied up the mpileup.[123].sam files to not use RNAME fields with no @SQ records.
I could have added the entries to the @SQ records, but this may have
caused more problems given we don't have the full chr17, just 4200
bytes of it and the mini-sequence in the ref.fa file.  Side-stepping
the whole CRAM and reference-required issue is the route of least
resistance.

Fixes #489