Use AC_SYS_LARGEFILE configure test instead of -D_FILE_OFFSET_BITS=64
Defines _FILE_OFFSET_BITS etc in config.h as appropriate.  On 64-bit
machines we don't need it, and it was just noise in make log output.
On 32-bit machines it now exactly parallels the settings applied in
HTSlib post samtools/htslib@f1327a703583bf61e2e7ffd11198487fee2a8422.

On 32-bit machines it is now *required* to run configure in order for
large files to be usable.  But it'll likely soon be non-optional for
everyone anyway.