Fixed bam_get_library() crash.
This was triggerable from samtools rmdup or samtools view -l when an
@RG header was present without an LB: record.

I also fixed the samtools view -l logic.  The man page states that it
only shows reads in that library, while the implementation was to only
show reads in that library or those where it could not determine the
library.

Fixes #538