Recreate missing @SQ header from BAM binary header
Some tools create BAM files with no text @SQ headers, which is unusual
but nonetheless valid.  Create a basic dummy header from the binary
header, as we may be merging with another file that does have @SQ
headers and having some but not all is not valid.

Fixes #548 and fixes #550.