Dynamically generate options
As was apparently mentioned at the meeting, it is probably better to
dynamically generate tab completions rather than to generate a one-off
script which the user then stores. This is what this new version does -
possible completions for subcommands and long-options are generated
on the fly, meaning that it is no longer necessary to try and keep the
completion file up-to-date. Additionally, this allows for support of
mulitple versions etc, by append lines to the bottom of the file, as
noted in the comments.

I have removed the previous version.