Fixed slight bug in default behaviour
Now handles files with spaces in their names correctly.

In doing so, I have basically made the change mentioned here:
https://github.com/samtools/samtools/pull/560#discussion-diff-59540631 I
have made the change in a slightly different location, however, as this
has the advantage of not allowing the user to expand '/' or similar when
they should be entering a subcommand (i.e.  when completing the first
argument).