Add `@CO` headers only once when merging sort temp files
Add MERGE_FIRST_CO for samtools sort (but don't expose it via the
merge command).  When set, bam_merge_core2() will copy `@CO` headers
from only the first file to be merged, so when a large sort produces
several temporary files to be merged, the headers won't be duplicated
in the final output.

This is a quick semi-hack to fix #563; there may be better ways to
fix this with greater refactoring.  See #390 for discussion of why
this simplistic approach is not suitable for samtools merge in general.