Write sort's temporary files alongside the *output* file
If several sorts are running simultaneously, the output filenames are
more likely to be distinct than the input filenames (which might often
be standard input).  This is a return to the pre-PR #441 behaviour.

If -T specifies a directory or the output is standard output, add a
random component to the prefix to try to avoid collisions.  (For output
to stdout with -T not specified, collisions would otherwise be certain.)

Fixes #523, fixes #535, et al.