Explicitly set @CURSES_LIB@ for --without-curses
If for some reason AX_WITH_CURSES is not expanded, autoconf will
never see CURSES_LIB/etc marked as output variables and not expand
@CURSES_LIB@/etc.  As AX_WITH_CURSES is within a conditional, this is
particularly confusing for --without-curses.  Fixes #528, but alters
nothing in normal circumstances.

(Set CURSES_LIB/etc to empty separately rather than with
AC_SUBST([CURSES_LIB], []), as with the latter the new code is
expanded to "else      fi", which is a shell syntax error.)