Add samtools view tests for extracting unmapped reads.
Add test to make samtools view search for region '*', i.e. unmapped reads.
Add an extra reference (ref3_unused) to view.001.sam, view.001.fa and
view.001.fa.fai.  This reference is the last in the header, but is not
used in the body of the file.

Detects the (htslib) bug described in issue samtools#568 (samtools
index/view bug with unmapped reads).