Depth/Mpileup -a fixes.  Fixes #374.
These work with and without regions (-r).  With bed files it's more
challenging.  They mostly work, but the distinction between -a and -aa
is blurred and bed regions that are entirely absent of data may be
omitted even with the -aa option.

It is possible to fix this, but it will need surgery on the bed
parsing code which these changes don't touch.  (Bed processing is
currently a completely opaque data structure with a minimal API.)