Added 30ish new tests for depth/mpileup -a and -aa.
Fixes samtools#374

Note that there is one case for each that is listed as a known
failure.  If we use -a with a BED file and the BED file specifies a
region which has zero coverage, but that reference has coverage
elsewhere (and we are not using -r to filter out that region), then
the output will still include zero data for the BED region.

Thus in this scenario -a is acting as -aa should.  Fixing this is
tricky, but probably there is little desire to use BED with the
inbetween -a (rather than -aa or no 'all' option).