Add libst.a so test suite can use print_error()/etc
Move print_error()/print_error_errno() to a new sam_utils.c so
that test programs can use them without getting samtools's main().
Build a new libst.a (but DON'T add it to "lib" or "install" targets,
as it's for internal use here in samtools only) linked against by
both samtools and the test programs, as this sort of shared code
is likely to increase.