Avoid strndup() and strsep() for portability reasons
strndup() is in POSIX but unavailable on e.g. OS X < 10.7;
strsep() is not a POSIX function.

No need to duplicate the input in confirm_rg() or get_first_rgid() as they
don't modify it, and the header has come from bam_hdr_dup() so hdr->text
is always NUL-terminated.