Samtools sort now multi-threads the decoding stage.
Also added an additional flag to bam_merge_core2 (MERGE_NO_TRANS) to
avoid a volumous amount of header merging, RG parsing and bam aux
tag hackery when called from sort.

This is a minor speed up, but more importantly it avoids reordering
header lines and moving of the per-read RG tag to the end of the
record.  Only implemented for sort; merge is unchanged.