Make samtools faidx write faster and detect write errors.
Replace the putchar() loop in faidx with a call to fwrite.  With glibc
the reduction in function call overhead gives a roughly 70% speed increase.

Add checks for write failure, so faidx is more likely to notice if
something goes wrong.  Also try to avoid a possible overflow if seq_len
is greater than INT_MAX - 60.