Output CRAM as temporary sort format iff there are more than 64k cigar ops.
Fixes #667.

In theory we could use CRAM for all temporary output files, but
currently it is too high in memory usage.  See the comments included
in this commit for a discussion on how this could be avoided, but for
now we are making the smallest change that fixes the issue as there
are larger issues yet to resolve in sort memory usage.