Make sort use all records read so far on each round.
bam_sort_core_ext() copies bam records into a fixed-size block
until one doesn't fit.  It then does a partial sort on the data
read so far.  Make it include the last record (that did not fit)
in the set that is sorted instead of carrying it over to the next
block.  It's easy to do and ensures sort will always make progress.

Also improve error handling in sort, mainly for memory allocation
failures.