Fix failure of bam2fq to write all reads on uncollated input
If the input to bam2fq is not collated by name, many reads will
look like they are single-ended because the paired reads may
not immediately follow each other.  Commit 33987e16 caused these
to be dropped if the -s flag was not used.  This restores the
previous behaviour where the records are written to the READ1 or
READ2 output file as appropriate.

Fixes #703