Make merge stable for name and by_tag ordering
Add a comparison on the position in the (concatenated) input files
to heap_lt() for name and by_tag orderings (there was one already
for position).  Ensures that a sort followed by a merge gives the
same ordering as a sort where no merge was needed.

This changes one of the test output files, as some reads that tied
before (and so came out in whatever order they happened to be on
the heap) are now ordered according to which input file they came
from.  This results in a few reads swapping in the output.  Both
orderings are valid according to the parameters given to merge.