Simplify lz4 build
We don't need to build an lz4 library, so move the lz4 build
rules into the main Makefile and remove the one in the lz4
sub-directory.  This is simpler and solves the problem of
propagating variables like CC to the recursive make.

Add some missing dependencies and fix up a couple of cases of
trailing white-space.