Improve collate interface and add tests
Add '-o' option to collate to specify exactly what output file
to write.  Makes writing alternate formats like sam easier.

Make the last command-line argument to collate optional.  If
present it is used to name the temporary files and the output
file if the -o and -O options have not been used.  If absent
the temporary files will go into /tmp (on UNIX-like systems)
or the result of GetTempPath() on Windows.

Include temporary files in the thread pool when writing.

Add tests for collate to test.pl.

Update manual page.