Make samtools idxstats work on SAM and CRAM files.
This isn't fast, but it fixes #303 and gives consistency across all
formats.  Cram is around 5x faster than unindexed BAM which in turn is
maybe 2x faster than unindexed SAM, but even CRAM is slow compared to
the almost instant speed of indexed BAM.

A possible future extension would be fixing the CRAM index format by
an additional two extra columns, but this is a spec related problem
and we'd still need to support older files.