Stats improvements
Computes the percentage of properly paired reads.

Includes '=' and 'X' CIGAR operators when counting mapped bases.

Implements overlap remove logic for paired-end reads in coverage
calculation and mapped base counting.  Overlap removal can be
turned on using the stats `-p` or `--remove-overlaps` option.

When a target file is in use calculates and prints the number of
bases inside the target, and the percentage of target bases with
coverage over a given threshold.  The coverage threshold can be
set using the `-g` or `--cov-threshold` option.

Adds more unit tests and augments some of the old ones.