Make samtools view -? work better
It -? appears on the command line, the long help message is printed
to stdout and the exit status is EXIT_SUCCESS.  If an unrecognised
option is used, the short message is printed to stderr and the
exit status is EXIT_FAILURE.

Using '-?' as a valid option along with getopt_long() is an unfortunate
choice as '-?' and an invalid long opt both cause getopt_long() to
return '?'.  If '?' is included in the list of options then optopt
ends up as '\0' for both cases too.

Resolve this by not including '?' in the list of valid options and
setting opterr = 0 to suppress the message that getopt_long()
would normally print.  The handler for '?' can now check optopt
to find out if the option really was '?', an unknown short
opt, or an unrecognised long opt.  For the error case, messages
similar to the ones output by getopt_long() are printed.  The
method used to print unrecognised options should hopefully
be portable between various getopt_long() implementations.  It
does mean international users will not see the translation of the
error message that they might normally get, though.