Increase default depth limit to 8000 and update man page
Set the default max-depth to 8000, which matches the limit that
was used by default for a single sample.  This also matches the
default used by the samtools depth command.

Add new #define MPLP_MAX_INDEL_DEPTH as it's now different to
max-depth.

Print the "Combined max depth is above 1M" warning based on the
number of files instead of the number of samples.  This matches
how the parameter is actually applied.

Update man page with the new default limit and add a note to
say that a minimum value is no longer enforced.