Add faidx option to annotate sequence names with strand
Long option only, useful mainly when writing the reverse
complement.  Allows choice of a simple '/rc' suffix on
reverse reads; '(+)' on forward and '(-)' on reverse which
matches the default output of "bedtools getfasta -s"; or a
custom option which allows pretty much any string.

Default is to leave the names unchanged.