Make /rc the default faidx/fqidx --mark-strand option
This allows forward strand names to be unchanged while making
it more obvious that a complemented sequence is from the opposite
strand.  If we allow requests for regions on different strands
(like `bedtools getfasta -s`) in the future then this will also
allow the same region to be output on both strands without
a name clash.

Update tests and documentation to match the new default.  Also
add --mark-strand documentation to the fqidx part of the man
page and add fqidx tests.