Improve plot-bamstats merge
Allow it to merge new LBC FBC LRL and FRL sections.

Remove GCD section.  It is not possible to merge this.

Preserve order of SN section to make comparing outputs easier.

Print merged GCC (and FBC, LBC) values with 2 decimal places to
better match the original stats output.

Warn rather than error on unprocessed segments, and ensure
the warning is printed for GCD and CHK.